Source/requirements.txt
aiogram == 2.25.2
python-dotenv == 1.0.1
firebase_admin == 6.5.0
httpx==0.27.0